KMP project structure for navigation-compose
