+++ navigation/navigation-compose/build.gradle
@@ -20,42 +20,57 @@ plugins {
     id("AndroidXPlugin")
     id("com.android.library")
     id("AndroidXComposePlugin")
-    id("org.jetbrains.kotlin.android")
+}
+
+androidXMultiplatform {
+    android()
+
+    sourceSets {
+        commonMain {
+            dependencies {
+                implementation(libs.kotlinStdlibCommon)
+                implementation(project(":compose:foundation:foundation-layout"))
+                api(project(":compose:animation:animation"))
+                api(project(":compose:runtime:runtime"))
+                api(project(":compose:runtime:runtime-saveable"))
+                api(project(":compose:ui:ui"))
+                api(project(":lifecycle:lifecycle-viewmodel-compose"))
+                api(project(":navigation:navigation-runtime"))
+            }
+        }
+
+        androidMain {
+            dependencies {
+                api("androidx.activity:activity-compose:1.8.0")
+            }
+        }
+
+        androidInstrumentedTest {
+            dependencies {
+                implementation(project(":compose:material:material"))
+                implementation project(":compose:test-utils")
+                implementation project(":compose:ui:ui-tooling")
+                implementation(project(":navigation:navigation-testing"))
+                implementation(project(":internal-testutils-navigation"))
+                implementation(project(":compose:ui:ui-test-junit4"))
+                implementation(project(":lifecycle:lifecycle-common"))
+                implementation(project(":lifecycle:lifecycle-common-java8"))
+                implementation(project(":lifecycle:lifecycle-livedata-core"))
+                implementation(project(":lifecycle:lifecycle-viewmodel"))
+                implementation(project(":lifecycle:lifecycle-viewmodel-savedstate"))
+                implementation(project(":activity:activity-ktx"))
+                implementation("androidx.collection:collection-ktx:1.1.0")
+                implementation(libs.testRunner)
+                implementation(libs.junit)
+                implementation(libs.truth)
+            }
+        }
+    }
 }
 
 dependencies {
-
-    implementation(libs.kotlinStdlib)
-    implementation(project(":compose:foundation:foundation-layout"))
-    api("androidx.activity:activity-compose:1.8.0")
-    api(project(":compose:animation:animation"))
-    api(project(":compose:runtime:runtime"))
-    api(project(":compose:runtime:runtime-saveable"))
-    api(project(":compose:ui:ui"))
-    api("androidx.lifecycle:lifecycle-viewmodel-compose:2.6.2")
-    api(projectOrArtifact(":navigation:navigation-runtime-ktx"))
-
-    androidTestImplementation(projectOrArtifact(":compose:material:material"))
-    androidTestImplementation project(":compose:test-utils")
-    androidTestImplementation projectOrArtifact(":compose:ui:ui-tooling")
-    androidTestImplementation(projectOrArtifact(":navigation:navigation-testing"))
-    androidTestImplementation(projectOrArtifact(":internal-testutils-navigation"), {
-        exclude group: "androidx.navigation", module: "navigation-common"
-    })
-    androidTestImplementation(projectOrArtifact(":compose:ui:ui-test-junit4"))
-    androidTestImplementation(projectOrArtifact(":lifecycle:lifecycle-common"))
-    androidTestImplementation(projectOrArtifact(":lifecycle:lifecycle-common-java8"))
-    androidTestImplementation(projectOrArtifact(":lifecycle:lifecycle-livedata-core"))
-    androidTestImplementation(projectOrArtifact(":lifecycle:lifecycle-viewmodel"))
-    androidTestImplementation(projectOrArtifact(":lifecycle:lifecycle-viewmodel-savedstate"))
-    androidTestImplementation(projectOrArtifact(":activity:activity-ktx"))
-    androidTestImplementation("androidx.collection:collection-ktx:1.1.0")
-    androidTestImplementation(libs.testRunner)
-    androidTestImplementation(libs.junit)
-    androidTestImplementation(libs.truth)
-
-    lintChecks(projectOrArtifact(":navigation:navigation-compose-lint"))
-    lintPublish(projectOrArtifact(":navigation:navigation-compose-lint"))
+    lintChecks(project(":navigation:navigation-compose-lint"))
+    lintPublish(project(":navigation:navigation-compose-lint"))
 }
 
 androidx {
@@ -64,7 +79,7 @@ androidx {
     inceptionYear = "2020"
     description = "Compose integration with Navigation"
     legacyDisableKotlinStrictApiMode = true
-    samples(projectOrArtifact(":navigation:navigation-compose:navigation-compose-samples"))
+    samples(project(":navigation:navigation-compose:navigation-compose-samples"))
 }
 
 android {
