+++ navigation/navigation-compose/src/androidInstrumentedTest/kotlin/androidx/navigation/compose/NavHostScreenShotTest.kt
