+++ navigation/navigation-compose/src/androidMain/kotlin/androidx/navigation/compose/NavBackStackEntryProvider.android.kt
@@ -17,44 +17,15 @@
 package androidx.navigation.compose
 
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.saveable.SaveableStateHolder
-import androidx.compose.ui.platform.LocalLifecycleOwner
-import androidx.compose.ui.platform.LocalSavedStateRegistryOwner
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
 import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavBackStackEntry
 import java.lang.ref.WeakReference
 import java.util.UUID
 
-/**
- * Provides [this] [NavBackStackEntry] as [LocalViewModelStoreOwner], [LocalLifecycleOwner] and
- * [LocalSavedStateRegistryOwner] to the [content] and saves the [content]'s saveable states with
- * the given [saveableStateHolder].
- *
- * @param saveableStateHolder The [SaveableStateHolder] that holds the saved states. The same
- * holder should be used for all [NavBackStackEntry]s in the encapsulating [Composable] and the
- * holder should be hoisted.
- * @param content The content [Composable]
- */
 @Composable
-public fun NavBackStackEntry.LocalOwnersProvider(
-    saveableStateHolder: SaveableStateHolder,
-    content: @Composable () -> Unit
-) {
-    CompositionLocalProvider(
-        LocalViewModelStoreOwner provides this,
-        LocalLifecycleOwner provides this,
-        LocalSavedStateRegistryOwner provides this
-    ) {
-        saveableStateHolder.SaveableStateProvider(content)
-    }
-}
-
-@Composable
-private fun SaveableStateHolder.SaveableStateProvider(content: @Composable () -> Unit) {
+internal actual fun SaveableStateHolder.SaveableStateProvider(content: @Composable () -> Unit) {
     val viewModel = viewModel<BackStackEntryIdViewModel>()
     // Stash a reference to the SaveableStateHolder in the ViewModel so that
     // it is available when the ViewModel is cleared, marking the permanent removal of this
